+++ public/index.php
@@ -3,6 +3,11 @@
 declare(strict_types=1);
 
 use Fhooe\Router\Router;
+use Monolog\Handler\StreamHandler;
+use Monolog\Logger;
+use Twig\Environment;
+use Twig\Loader\FilesystemLoader;
+use Twig\TwigFunction;
 
 require "../vendor/autoload.php";
 
@@ -26,12 +31,28 @@ if (DEBUG) {
 // Create a new Router object.
 $router = new Router();
 
+// Create a monolog instance for logging in the skeleton. Pass it to the router to receive its log messages too.
+$logger = new Logger("skeleton-logger");
+$logger->pushHandler(new StreamHandler(__DIR__ . "/../logs/router.log"));
+$router->setLogger($logger);
+
+// Create a new twig instance for advanced templates.
+$twig = new Environment(
+    new FilesystemLoader("../views"),
+    [
+        "cache" => "../cache",
+        "auto_reload" => true
+    ]
+);
+$twig->addFunction(new TwigFunction("url_for", [Router::class, "urlFor"]));
+$twig->addFunction(new TwigFunction("get_base_path", [Router::class, "getBasePath"]));
+
 // Set a base path if your code is not in your server's document root.
 $router->setBasePath("/code/fhooe-router-skeleton/public");
 
 // Set a 404 callback that is executed when no route matches.
 $router->set404Callback(function () {
-    require __DIR__ . "/../views/404.html";
+    require __DIR__ . "/../views/404.php";
 });
 
 // Define your routes with the get() and post() methods.
@@ -43,8 +64,16 @@ $router->get("/form", function () {
     require __DIR__ . "/../views/form.php";
 });
 
-$router->post("/form", function () {
-    require __DIR__ . "/../views/form.php";
+$router->post("/formresult", function () {
+    require __DIR__ . "/../views/formresult.php";
+});
+
+$router->get("/twigform", function () use ($twig) {
+    $twig->display("twigform.html.twig");
+});
+
+$router->post("/twigform", function () use ($twig) {
+    $twig->display("twigform.html.twig", ["emailInput" => $_POST["emailInput"]]);
 });
 
 // Run the router to get the party started.
@@ -64,6 +93,6 @@ switch ($route) {
         require __DIR__ . "/../views/form.php";
         break;
     default:
-        require __DIR__ . "/../views/404.html";
+        require __DIR__ . "/../views/404.php";
         break;
 }*/